+++ test/reviewer/setup/generator_test.rb
@@ -83,23 +83,12 @@ module Reviewer
         assert_match(/^# Review Ruby syntax/, yaml)
       end
 
-      def test_generates_js_tools
-        generator = Generator.new([:eslint])
-        yaml = generator.generate
-
-        parsed = YAML.safe_load(yaml)
-        assert parsed.key?('eslint')
-        assert_match(/eslint/, parsed['eslint']['commands']['review'])
-      end
-
       def test_uses_yarn_when_lockfile_present
         Dir.mktmpdir do |dir|
           Pathname(dir).join('yarn.lock').write('')
 
-          generator = Generator.new([:eslint], project_dir: Pathname(dir))
-          yaml = generator.generate
+          parsed = YAML.safe_load(Generator.new([:eslint], project_dir: Pathname(dir)).generate)
 
-          parsed = YAML.safe_load(yaml)
           assert_match(/^yarn /, parsed['eslint']['commands']['review'])
         end
       end
@@ -108,20 +97,16 @@ module Reviewer
         Dir.mktmpdir do |dir|
           Pathname(dir).join('pnpm-lock.yaml').write('')
 
-          generator = Generator.new([:eslint], project_dir: Pathname(dir))
-          yaml = generator.generate
+          parsed = YAML.safe_load(Generator.new([:eslint], project_dir: Pathname(dir)).generate)
 
-          parsed = YAML.safe_load(yaml)
           assert_match(/^pnpm exec /, parsed['eslint']['commands']['review'])
         end
       end
 
       def test_defaults_to_npx_without_lockfile
         Dir.mktmpdir do |dir|
-          generator = Generator.new([:eslint], project_dir: Pathname(dir))
-          yaml = generator.generate
+          parsed = YAML.safe_load(Generator.new([:eslint], project_dir: Pathname(dir)).generate)
 
-          parsed = YAML.safe_load(yaml)
           assert_match(/^npx /, parsed['eslint']['commands']['review'])
         end
       end
@@ -129,11 +114,8 @@ module Reviewer
       def test_does_not_alter_ruby_commands
         Dir.mktmpdir do |dir|
           Pathname(dir).join('yarn.lock').write('')
+          parsed = YAML.safe_load(Generator.new([:rubocop], project_dir: Pathname(dir)).generate)
 
-          generator = Generator.new([:rubocop], project_dir: Pathname(dir))
-          yaml = generator.generate
-
-          parsed = YAML.safe_load(yaml)
           assert_match(/^bundle exec/, parsed['rubocop']['commands']['review'])
         end
       end